Migrate Mutation.updateSchemaPolicyForProject
