+++ packages/services/api/src/modules/policy/index.ts
@@ -1,7 +1,7 @@
 import { createModule } from 'graphql-modules';
 import { SchemaPolicyApiProvider } from './providers/schema-policy-api.provider';
 import { SchemaPolicyProvider } from './providers/schema-policy.provider';
-import { resolvers } from './resolvers';
+import { resolvers } from './resolvers.generated';
 import typeDefs from './module.graphql';
 
 export const schemaPolicyModule = createModule({
